Modyfying the whole project onto BigDecimal.
In order to obtain a mathematical precision, I modified all of the classes from double types onto the BigDecimal class objects.

+++ src/ResetCalc.java
@@ -1,7 +1,9 @@
+import java.math.BigDecimal;
+
 public class ResetCalc {
     public void reset(){
-        Program.firstNumber = 0;
-        Program.secondNumber = 0;
+        Program.firstNumber = BigDecimal.valueOf(0);
+        Program.secondNumber = BigDecimal.valueOf(0);
         System.out.println(Program.firstNumber);
     }
 }
